AC_Sprayer: explicitly cast parameter to move_servo to fix conversion error

+++ libraries/AC_Sprayer/AC_Sprayer.cpp
@@ -182,7 +182,7 @@ void AC_Sprayer::update()
         float pos = ground_speed * _pump_pct_1ms;
         pos = MAX(pos, 100 *_pump_min_pct); // ensure min pump speed
         pos = MIN(pos,10000); // clamp to range
-        SRV_Channels::move_servo(SRV_Channel::k_sprayer_pump, pos, 0, 10000);
+        SRV_Channels::move_servo(SRV_Channel::k_sprayer_pump, uint16_t(pos), 0, 10000);
         SRV_Channels::set_output_pwm(SRV_Channel::k_sprayer_spinner, _spinner_pwm);
         _flags.spraying = true;
     } else {
